+++ docs/index.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <meta charset='utf-8'>
 <title>reflexbox</title>
-<meta name='viewport' content='width=640, initial-scale=1'>
+<meta name='viewport' content='width=640, initial-scale=0.5'>
 <style>
 body {
   font-family: -apple-system, sans-serif;
